Adds test-unit to Gemfile
When first cloning this repository and executing the following:

```
bundle install
bundle exec rake test
```

the user is met with an error related to `test/unit` being a file
that does not exist during the require at the top of the first test
file.  Adding this to the Gemfile and rebundling resolves the issue.